refactor to remove hittable abstraction, maybe 15-20% faster but maybe not and not the big boost i hoped for

+++ rtweekend.h
@@ -14,12 +14,12 @@ double degrees_to_radians(double degrees) {
   return degrees * pi / 180.0;
 }
 
-double random_double() {
-  return (float)rand() / ((float)(RAND_MAX) + 1);
+float random_double() {
+  return rand() / (RAND_MAX + 1.0);
 }
 
-double random_double_range(double min, double max) {
-  return min + (max - min) * (float)rand() / ((float)(RAND_MAX) + 1);
+float random_double_range(float min, float max) {
+  return min + (max - min) * rand() / (RAND_MAX + 1.0);
 }
 
 //double random_double() {
